chore: bundle 관련 api, type src/entities/bundle로 이동 (#206)
