+++ src/entities/bundle/constants/bundle.ts
@@ -1,1 +1,8 @@
-export const TIME_SLOTS = ['MORNING', 'AFTERNOON', 'EVENING', 'DAWN'] as const
+import type { TimeSlot } from '@/shared/types/common'
+
+export const TIME_SLOTS = [
+  'MORNING',
+  'AFTERNOON',
+  'EVENING',
+  'DAWN',
+] as const satisfies readonly TimeSlot[]
