+++ src/entities/playlist/types/playlist.ts
@@ -1,5 +1,4 @@
 import type { ShareCode } from '@/features/auth'
-import type { TimeSlot } from '@/shared/types'
 import type { SortType } from '@/shared/ui/ContentHeader'
 
 export type FEED_CD_LIST_TAB_TYPE = 'cds' | 'likes'
@@ -128,11 +127,4 @@ export interface Playlist {
   cdResponse?: {
     cdItems: CdCustomData[]
   }
-}
-
-export interface BundleInfo {
-  bundleId: number
-  title: string
-  timeSlot: TimeSlot
-  playlists: Playlist[]
 }
