+++ src/pages/curation/index.tsx
@@ -2,10 +2,10 @@ import { useNavigate, useOutletContext } from 'react-router-dom'
 
 import styled from 'styled-components'
 
-import { useToast } from '@/app/providers'
 import { Share, StartBlack } from '@/assets/icons'
-import { usePlaylistDetails, type BundleInfo } from '@/entities/playlist'
-import { useCopyShareUrl, useSingleSelect } from '@/shared/lib'
+import type { BundleInfo } from '@/entities/bundle'
+import { usePlaylistDetails } from '@/entities/playlist'
+import { useShareLink, useSingleSelect } from '@/shared/lib'
 import { flexRowCenter } from '@/shared/styles/mixins'
 import { Cd, ContentHeader, Divider } from '@/shared/ui'
 import type { SortType } from '@/shared/ui/ContentHeader'
@@ -14,36 +14,20 @@ import { SearchResultItem } from '@/widgets/search'
 const Curation = () => {
   const bundle = useOutletContext<BundleInfo>()
   const navigate = useNavigate()
-  const { toast } = useToast()
-  const { copyShareUrl } = useCopyShareUrl()
+  const { shareLink } = useShareLink()
   const { selected, onSelect } = useSingleSelect<SortType>('POPULAR')
 
   const ids = bundle.playlists.map((p) => p.playlistId)
 
   const { data } = usePlaylistDetails(ids)
 
-  const handleShare = async () => {
-    // 브라우저 지원 여부 및 https 체크 (미지원 시 함수로 별도 copy 처리)
-    if (typeof window === 'undefined' || !navigator.share) {
-      copyShareUrl('curation', bundle.bundleId)
-      return
-    }
-
-    // 공유할 데이터 설정
-    const shareData = {
+  const handleShare = () => {
+    shareLink({
+      copyType: 'curation',
+      copyValue: bundle.bundleId,
       title: `[DEULAK] ${bundle.title}`,
       url: `${window.location.origin}/curation/${bundle.bundleId}`,
-    }
-
-    try {
-      await navigator.share(shareData)
-      toast('LINK')
-    } catch (error) {
-      // 사용자가 공유를 취소한 경우 외의 에러 케이스
-      if ((error as Error).name !== 'AbortError') {
-        console.error('URL 공유 중 에러 발생: ', error)
-      }
-    }
+    })
   }
 
   return (
